Extract some components to their own file
